add qeo

+++ RC2026_H7/RC26_Motor/Src/RC_dji_motor.cpp
@@ -115,7 +115,7 @@ namespace motor
 		current 	= (float)(int16_t)(((uint16_t)rx_data[4] << 8) | (uint16_t)rx_data[5]);
 		temperature = (float)(int8_t)rx_data[6];
 		
-		angle 		= (float)encoder * (1.f / 8192.f) * TWO_PI;
+		angle 		= (float)encoder * (float)(1.0 / 8192.0) * TWO_PI;
 		
 		// 计算转子旋转圈数
 		if (can_rx_is_first != true)
